finish splits
