+++ jest.config.json
@@ -6,7 +6,14 @@
 	"verbose": true,
 	"roots": ["<rootDir>/test", "<rootDir>/src"],
 	"collectCoverageFrom": ["src/**/*.ts"],
-	"coveragePathIgnorePatterns": ["src/index.ts"],
+	"coveragePathIgnorePatterns": [
+    "src/index.ts",
+    "src/browser.ts",
+    "src/node.ts",
+    "src/controllers/index.ts",
+    "src/models/index.ts",
+    "src/views/index.ts"
+  ],
 	"coverageThreshold": {
 		"global": {
 			"branches": 50,
