+++ package.json
@@ -1,7 +1,11 @@
 {
   "name": "@tinystacks/ops-core",
   "version": "0.3.2",
-  "main": "dist/index.js",
+  "main": "./dist/index.js",
+  "exports": {
+    "node": "./dist/node.js",
+    "browser": "./dist/browser.js"
+  },
   "type": "module",
   "repository": {
     "type": "git",
@@ -35,6 +39,7 @@
     "@babel/cli": "^7.20.7",
     "@babel/core": "^7.20.12",
     "@babel/preset-env": "^7.20.2",
+    "@types/http-status-codes": "^1.2.0",
     "@types/jest": "^29.4.0",
     "@types/lodash.get": "^4.4.7",
     "@types/lodash.isnil": "^4.0.7",
@@ -56,6 +61,7 @@
   "dependencies": {
     "@tinystacks/ops-model": "^0.4.0",
     "@types/react": "^18.0.28",
+    "http-status-codes": "^2.2.0",
     "lodash.get": "^4.4.2",
     "lodash.isnil": "^4.0.0"
   }
