+++ src/node.ts
@@ -1,0 +1,3 @@
+export * as Controllers from './controllers/index.js';
+export * as Models from './models/index.js';
+export * as Types from './types.js';
