+++ src/types.ts
@@ -5,8 +5,3 @@ export type Json = {
 export type Typed = Json & {
   type?: string | undefined;
 }
-
-export interface Parsable<T extends Typed, U extends T> {
-  fromJson (object: T, ...args: any[]): Promise<U> | U;
-  toJson (): Promise<T> | T;
-}
