+++ src/views/index.ts
@@ -1,0 +1,2 @@
+export * from './base-widget.js';
+export * from './widget.js';
